requirements.txt
torch>=2.0.0
diffusers[torch]>=0.24.0
transformers>=4.36.0
accelerate>=0.26.0
peft>=0.8.0
datasets>=2.15.0
Pillow>=10.0.0
pyyaml>=6.0
opencv-python>=4.8.0
numpy>=1.24.0
fastapi
uvicorn
requests
python-multipart